Made changes so TORQUE would compile on Solaris5 platform.
git-svn-id: svn://clusterresources.com/torque/branches/2.3-fixes@3390 8f940c70-5916-0410-ac46-d1fa2fa6ea29

+++ src/resmom/catch_child.c
@@ -1380,7 +1380,7 @@ static void preobit_reply(
     if (LOGLEVEL >= 7)
       {
       snprintf(log_buffer, 1024, "epilog subtask created with pid %d - substate set to JOB_SUBSTATE_OBIT - registered post_epilogue",
-               cpid);
+               (int)cpid);
 
       log_record(
         PBSEVENT_DEBUG,
