+++ src/server/pbsd_init.c
@@ -1646,7 +1646,7 @@ static void catch_child(
         msg_daemonname,
         log_buffer);
 #else
-      DBPRT(("catch_child caught pid %d\n", pid));
+      DBPRT(("catch_child caught pid %d\n", (int)pid));
 #endif
       }
 
@@ -1677,7 +1677,7 @@ static void catch_child(
             msg_daemonname,
             log_buffer);
 #else
-          DBPRT(("catch_child found work task found for pid %d\n", pid));
+          DBPRT(("catch_child found work task found for pid %d\n", (int)pid));
 #endif
           }
 
@@ -1694,7 +1694,7 @@ static void catch_child(
 
       log_err(-1, "catch_child", log_buffer);
 #else
-      DBPRT(("catch_child no work task found for pid %d\n", pid));
+      DBPRT(("catch_child no work task found for pid %d\n", (int)pid));
 #endif
       }
     }    /* END while (1) */
